extract aggregate
